+++ GameServer/quests/Quests/QuestScripts/albion/TombWithAView.cs
@@ -267,7 +267,7 @@ namespace DOL.GS.Quests.Albion
                     log.Warn("Could not find " + LadyGrynoch.Name + ", creating her ...");
                 //k109: My preference, no guildname for quest NPCs.  Uncomment if you like that...
                 //LadyGrynoch.GuildName = "Part of " + questTitle + " Quest";
-                LadyGrynoch.Realm = (byte)eRealm.Albion;
+                LadyGrynoch.Realm = eRealm.Albion;
                 LadyGrynoch.CurrentRegionID = 1;
 
                 GameNpcInventoryTemplate template = new GameNpcInventoryTemplate();
